Created separate VO objects for different rest end-points
