+++ sawtooth-asset-app/src/main/java/com/mycompany/blockchain/sawtooth/app/controller/WalletController.java
@@ -3,7 +3,6 @@ package com.mycompany.blockchain.sawtooth.app.controller;
 import java.io.UnsupportedEncodingException;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.json.JsonParserFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,14 +10,12 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.fasterxml.jackson.core.JsonParser;
 import com.google.protobuf.ByteString;
 import com.google.protobuf.InvalidProtocolBufferException;
 import com.googlecode.protobuf.format.JsonFormat;
 import com.mycompany.blockchain.sawtooth.app.service.WalletService;
-import com.mycompany.blockchain.sawtooth.app.vo.AssetVO;
-import com.mycompany.blockchain.sawtooth.app.vo.WalletVO;
-import com.mycompany.blockchain.sawtooth.client.asset.AssetPayloadClientService;
+import com.mycompany.blockchain.sawtooth.app.vo.wallet.WalletTransferVO;
+import com.mycompany.blockchain.sawtooth.app.vo.wallet.WalletVO;
 import com.mycompany.blockchain.sawtooth.client.wallet.WalletPayloadClientService;
 import com.mycompany.blockchain.sawtooth.wallet.protobuf.SawtoothWalletPayload;
 import com.mycompany.blockchain.sawtooth.wallet.protobuf.SawtoothWalletPayload.CreateWallet;
@@ -28,7 +25,6 @@ import com.mycompany.blockchain.sawtooth.wallet.protobuf.SawtoothWalletPayload.T
 import com.mycompany.blockchain.sawtooth.wallet.protobuf.SawtoothWalletPayload.Withdraw;
 import com.mycompany.blockchain.sawtooth.wallet.protobuf.Wallet;
 
-import lombok.extern.java.Log;
 import lombok.extern.slf4j.Slf4j;
 import sawtooth.sdk.processor.exceptions.ValidatorConnectionError;
 import sawtooth.sdk.protobuf.ClientBatchSubmitResponse.Status;
@@ -102,7 +98,7 @@ public class WalletController {
 	}
 
 	@RequestMapping(consumes = "application/json", method = RequestMethod.POST, path = "/wallet/transfer")
-	public @ResponseBody int transferWallet(@RequestBody WalletVO walletVo) throws Exception {
+	public @ResponseBody int transferWallet(@RequestBody WalletTransferVO walletVo) throws Exception {
 		TransferPayment transferPayment = TransferPayment.newBuilder()
 				.setSourceCustomerId(walletVo.getCustomerId()).setAmount(walletVo.getBalance())
 				.setDestCustomerId(walletVo.getDestCustomerId()).build();
