+++ sawtooth-asset-app/src/main/java/com/mycompany/blockchain/sawtooth/app/vo/loan/CreateLoanVO.java
@@ -1,0 +1,24 @@
+/**
+ * 
+ */
+package com.mycompany.blockchain.sawtooth.app.vo.loan;
+
+import lombok.Data;
+
+/**
+ * Value Object for the UI, UI can have its own schema mapped to the LoanVO that can be then
+ * converted to Loan
+ * 
+ * @author Sandip Nirmal
+ *
+ */
+
+@Data
+public class CreateLoanVO {
+	String lenderId;
+	String borrowerId;
+	String assetId;
+	float roi;
+	int requestedAmount;
+}
+
